backend/requirements.txt
Django>=6.0.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.3.1
python-decouple>=3.8
django-ratelimit>=4.1.0
psycopg2-binary>=2.9.0
